requirements.txt
asgiref==3.8.1
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
folium==0.19.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.2.3
xyzservices==2024.9.0
